moving shared code to its own folder
