+++ OpensslPkg/Library/OpensslLib/OpenssLibShared.inf
@@ -1,84 +1,41 @@
 ## @file
-#  This module provides OpenSSL Binary implementation.
+#  This module provides OpenSSL Library implementation with ECC and TLS
+#  features.
+#
+#  This library should be used if a module module needs ECC in TLS, or
+#  asymmetric cryptography services such as X509 certificate or PEM format
+#  data processing.  This library increases the size overhead up to ~115 KB
+#  compared to OpensslLib.inf library instance.
+#
+#  Copyright (c) 2010 - 2020, Intel Corporation. All rights reserved.<BR>
+#  (C) Copyright 2020 Hewlett Packard Enterprise Development LP<BR>
+#  SPDX-License-Identifier: BSD-2-Clause-Patent
 #
 ##
 
 [Defines]
   INF_VERSION                    = 0x00010005
-  BASE_NAME                      = SharedCryptoBin
-  FILE_GUID                      = 76ABA88D-9D16-49A2-AA3A-DB6112FAC5CB
-  MODULE_TYPE                    = UEFI_APPLICATION
+  BASE_NAME                      = OpensslLibFull
+  MODULE_UNI_FILE                = OpensslLibFull.uni
+  FILE_GUID                      = AB9E2231-D8FC-433F-9F27-FA293B64FB2C
+  MODULE_TYPE                    = BASE
   VERSION_STRING                 = 1.0
-  ENTRY_POINT                    = UefiMain
+  LIBRARY_CLASS                  = OpensslLib
+  CONSTRUCTOR                    = OpensslLibConstructor
 
   DEFINE OPENSSL_PATH            = openssl
   DEFINE OPENSSL_FLAGS           = -DL_ENDIAN -DOPENSSL_SMALL_FOOTPRINT -D_CRT_SECURE_NO_DEPRECATE -D_CRT_NONSTDC_NO_DEPRECATE -DOPENSSL_NO_ASM
   DEFINE OPENSSL_FLAGS_CONFIG    =
-#
-# The following information is for reference only and not required by the build tools.
+
 #
 #  VALID_ARCHITECTURES           = IA32 X64 ARM AARCH64
 #
 
-[LibraryClasses]
-  UefiApplicationEntryPoint
-  IntrinsicLib
-  BaseLib
-
-[Packages]
-  MdePkg/MdePkg.dec
-  MdeModulePkg/MdeModulePkg.dec
-  OpensslPkg/OpensslPkg.dec
-
 [Sources]
-  Shared/InternalCryptLib.h
-  SharedCryptoBin.c
-  Shared/SharedOpenssl.c
-  Shared/SharedOpenssl.h
-  Shared/SharedCryptoProtocol.h
-  Shared/SharedCryptDecl.h
-  Shared/SharedCryptoCrt.c
-  Shared/Syscall/CrtWrapper.c
-  Shared/Syscall/BaseMemAllocation.c
-  Shared/Syscall/TimerWrapper.c
-  Shared/Math/DivS64x64Remainder.c
-  # Big Number Crypto API
-  Shared/BN/CryptBn.c
-  # Cipher Crypto API
-  Shared/Cipher/CryptAeadAesGcm.c
-  Shared/Cipher/CryptAes.c
-
-  # Hash Crypto API
-  Shared/Hash/CryptMd5.c
-  Shared/Hash/CryptSha1.c
-  Shared/Hash/CryptSha256.c
-  Shared/Hash/CryptSha512.c
-  Shared/Hash/CryptSm3.c
-
-  # HMAC API
-  Shared/Hmac/CryptHmac.c
-
-  # KDF API
-  Shared/Kdf/CryptHkdf.c
-
-  # PEM API
-  Shared/Pem/CryptPem.c
-
-  # PK API
-  Shared/Pk/CryptAuthenticode.c
-  Shared/Pk/CryptPkcs7VerifyCommon.c
-  Shared/Pk/CryptDh.c
-  Shared/Pk/CryptPkcs5Pbkdf2.c
-  Shared/Pk/CryptPkcs1Oaep.c
-
-  # Random Crypto API
-  rand_pool.c
-  Shared/Rand/CryptRand.h
-  Shared/Rand/CryptRand.c
-
-  # OpenSSL Crypto Library
+  OpensslLibConstructor.c
   $(OPENSSL_PATH)/e_os.h
   $(OPENSSL_PATH)/ms/uplink.h
+# Autogenerated files list starts here
   $(OPENSSL_PATH)/crypto/aes/aes_cbc.c
   $(OPENSSL_PATH)/crypto/aes/aes_cfb.c
   $(OPENSSL_PATH)/crypto/aes/aes_core.c
@@ -673,12 +630,22 @@
   $(OPENSSL_PATH)/crypto/ec/curve448/word.h
   $(OPENSSL_PATH)/crypto/ec/curve448/arch_32/arch_intrinsics.h
   $(OPENSSL_PATH)/crypto/ec/curve448/arch_32/f_impl.h
+# Autogenerated files list ends here
   buildinf.h
   ossl_store.c
+  rand_pool.c
+#  SslNull.c
+#  EcSm2Null.c
+
+[Packages]
+  MdePkg/MdePkg.dec
+  CryptoPkg/CryptoPkg.dec
+  OpensslPkg/OpensslPkg.dec
+
+[LibraryClasses]
+
 
 [BuildOptions]
-  MSFT:*_*_*_DLINK_FLAGS  = /DLL /SUBSYSTEM:CONSOLE /VERSION:1.0
-  MSFT:*_*_*_GENFW_FLAGS = --keepoptionalheader
   #
   # Disables the following Visual Studio compiler warnings brought by openssl source,
   # so we do not break the build with /WX option:
